The demo is just an hello world, rename
A real demo (with app compat as a dependency) will be added later.
