+++ apps/helloworld/proguard-rules.pro
