+++ apps/helloworld/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
-          package="com.microsoft.sonoma.demo">
+          package="com.microsoft.sonoma.helloworld">
 
     <application
         android:allowBackup="true"
@@ -9,7 +9,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
-        <activity android:name="com.microsoft.sonoma.demo.MainActivity">
+        <activity android:name=".HelloWorldActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
