+++ apps/helloworld/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
                 android:paddingLeft="@dimen/activity_horizontal_margin"
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
-                tools:context="com.microsoft.sonoma.demo.MainActivity">
+                tools:context=".HelloWorldActivity">
 
     <TextView
         android:layout_width="wrap_content"
