+++ settings.gradle
@@ -10,4 +10,4 @@ include ':test'
 // test apps
 include ':apps'
 include ':apps:sasquatch'
-include ':apps:demo'
+include ':apps:helloworld'
